backend/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13.0"
fastapi = ">=0.109.1,<1"
requests = "^2.32.4"
types-requests = "^2.31.0"
pydantic = "^2.10.2"
pyhumps = "^3.8.0"
uvicorn = ">=0.23.1,<1"
python-ulid = "^1.1.0"
boto3 = "^1.37.0"
python-jose = "^3.4.0"
pg8000 = "^1.30.3"
argparse = "^1.4.0"
tenacity = "<=8.3.0"
langdetect = "^1.0.9"
types-retry = ">=0.9.9.4,<1"
opensearch-py = ">=2.0.0"
requests-aws4auth = ">=1.0.0"
duckduckgo-search = "^7.3.0"
boto3-stubs = {extras = ["bedrock", "bedrock-agent-runtime", "bedrock-runtime", "boto3"], version = "^1.37.0"}
firecrawl-py = "^1.11.1"
reretry = "^0.11.8"
python-pptx = "^1.0.2"
python-docx = "^1.1.2"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
